Seite erstellen Modal teilweise ohne Language select

+++ protected/widgets/menu/NLevelNav.php
@@ -126,17 +126,24 @@ EOP;
 					if(! array_key_exists('url', $item))
 						$item['url'] = '#';
 					
-					if(! array_key_exists('items', $item) || (count($item) <= 0))
-						echo CHtml::link($item['label'], $item['url']);
-					else 
+					if(array_key_exists('ajax', $item))
 					{
-						if($submenu)
-							echo CHtml::link($item['label'], '#', array('class'=>'trigger right-caret', 'data-toggle'=>'dropdown'));
-						else
-							echo CHtml::link($item['label'].'<span class="caret"></span>', '#', array('class'=>'dropdown-toggle', 'data-toggle'=>'dropdown'));
-						
-						if(count($item['items']) > 0)
-							$this->buildNav($item['items'], 'dropdown-menu sub-menu', true);
+						echo CHtml::link($item['label'], '', array('onclick' => $item['ajax']));
+					}
+					else
+					{
+						if(! array_key_exists('items', $item) || (count($item) <= 0))
+							echo CHtml::link($item['label'], $item['url']);
+						else 
+						{
+							if($submenu)
+								echo CHtml::link($item['label'], '#', array('class'=>'trigger right-caret', 'data-toggle'=>'dropdown'));
+							else
+								echo CHtml::link($item['label'].'<span class="caret"></span>', '#', array('class'=>'dropdown-toggle', 'data-toggle'=>'dropdown'));
+							
+							if(count($item['items']) > 0)
+								$this->buildNav($item['items'], 'dropdown-menu sub-menu', true);
+						}
 					}
 					echo '</li>';
 				}
